// app/services/types/home.ts
export interface CreateAddressRequest {
  fullName: string;
  user_id: number;
  address: string;
  lat: string;
  long: string;
  type: string;
}